refactor: jdbc 공통 로직 분리 - try with resource 구문 활용 - 람다로 쿼리 별 다른 로직을 수행하도록 구현
