+++ src/main/java/nextstep/domain/repository/executor/PrepareStatementExecutor.java
@@ -1,0 +1,11 @@
+package nextstep.domain.repository.executor;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface PrepareStatementExecutor<T> {
+
+    T execute(PreparedStatement pstmt) throws SQLException;
+
+}
