+++ src/main/java/nextstep/domain/repository/executor/PrepareStatementResultSetExecutor.java
@@ -1,0 +1,12 @@
+package nextstep.domain.repository.executor;
+
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface PrepareStatementResultSetExecutor<T> {
+
+    T execute(PreparedStatement pstmt, ResultSet rs) throws SQLException;
+
+}
